+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## Unreleased
+
+### Updated
+1. Fixed listen event decoding (cf. https://github.com/uhppoted/uhppoted-python/issues/3)
+
 ## [0.8.7](https://github.com/uhppoted/uhppoted-python/releases/tag/v0.8.7) - 2023-12-01
 
 ### Added